+++ src/Products/Five/security.py
@@ -34,6 +34,12 @@ from zope.security.simplepolicies import ParanoidSecurityPolicy
 
 from AccessControl.SecurityInfo import ClassSecurityInfo
 from AccessControl.SecurityManagement import getSecurityManager
+from AccessControl.Permission import _registeredPermissions
+from AccessControl.Permission import pname
+
+import Products
+
+from Globals import ApplicationDefaultPermissions
 
 CheckerPublicId = 'zope.Public'
 CheckerPrivateId = 'zope2.Private'
@@ -155,3 +161,21 @@ def protectClass(klass, permission_id):
         # Zope 2 uses string, not unicode yet
         perm = str(permission.title)
         security.declareObjectProtected(perm)
+
+def create_permission_from_permission_directive(permission, event):
+    """When a new IPermission utility is registered (via the <permission />
+    directive), create the equivalent Zope2 style permission.
+    """
+    
+    global _registeredPermissions
+    
+    zope2_permission = permission.title
+    roles = ('Manager',)
+    
+    if not _registeredPermissions.has_key(zope2_permission):
+        _registeredPermissions[zope2_permission] = 1
+        
+        Products.__ac_permissions__ += ((zope2_permission, (), roles,),)
+        
+        mangled = pname(zope2_permission)
+        setattr(ApplicationDefaultPermissions, mangled, roles)